fix: CompletedCourse를 불변 객체로 변경
- Setter 삭제
- 모든 필드에 final 속성 추가
